pad stuffo and more stuff stuff
ah pure bliss #2

it's possible now to write a full interactive Game of Life implementation
in 20-something lines of code

+++ src/canvas/generic.rs
@@ -144,7 +144,7 @@ impl<Spec: DeviceSpec> crate::Canvas for DeviceCanvas<'_, Spec> {
 			let old = button.get_old(self).quantize(Spec::COLOR_PRECISION);
 			let new = button.get(self).quantize(Spec::COLOR_PRECISION);
 			if new != old {
-				changes.push((button.x(), button.y(), new));
+				changes.push((button.x as u32, button.y as u32, new));
 			}
 		}
